normalize inflection from indexes to indices
