+++ app/models/search/indices/documents_index.rb
@@ -1,5 +1,5 @@
 module Search
-  module Indexes
+  module Indices
 
     class DocumentsIndex < Index
       def mappings
