+++ app/models/search/indices/language_index.rb
@@ -1,10 +1,10 @@
 module Search
-  module Indexes
+  module Indices
 
-    class IdentityIndex < Index
+    class LanguageIndex < Index
       def mappings
         {
-          identity: {
+          language: {
             properties: {
               id: {type: 'string', index: 'not_analyzed'},
               name: {
@@ -13,6 +13,15 @@ module Search
                   name:    {type: 'string', index: 'not_analyzed'},
                   full:    {type: 'string', analyzer: 'full_str'},
                   partial: {type: 'string', analyzer: 'partial_str'}
+                }
+              },
+              full_name: {
+                type: 'multi_field',
+                fields: {
+                  full_name: {type: 'string', index: 'not_analyzed'},
+                  full:      {type: 'string', analyzer: 'full_str'},
+                  partial:   {type: 'string', analyzer: 'partial_str'}
+
                 }
               }
             }
