+++ app/models/search/indices/subject_index.rb
@@ -1,5 +1,5 @@
 module Search
-  module Indexes
+  module Indices
 
     class SubjectIndex < Index
       def mappings
