+++ test/models/search/indices/identity_index_test.rb
@@ -2,12 +2,12 @@ require 'securerandom'
 require 'test_helper'
 
 module Search
-  module Indexes
-    class GradeIndexTest < ActiveSupport::TestCase
+  module Indices
+    class IdentityIndexTest < ActiveSupport::TestCase
       @@initial_setup = true
 
       setup do
-        index = GradeIndex.new
+        index = IdentityIndex.new
         if @@initial_setup || !index.index_exists?
           index.reset_index!
           @@initial_setup = false
@@ -15,17 +15,17 @@ module Search
       end
 
       test "#type_name" do
-        index = GradeIndex.new
-        assert_equal 'grade', index.type_name
+        index = IdentityIndex.new
+        assert_equal 'identity', index.type_name
       end
 
       test "#index_name is composed with type and environment" do
-        index = GradeIndex.new
-        assert_equal 'grades__test', index.index_name
+        index = IdentityIndex.new
+        assert_equal 'identities__test', index.index_name
       end
 
       test '#create_index! creates a new index' do
-        index = GradeIndex.new
+        index = IdentityIndex.new
         index.delete_index!
         refute index.index_exists?
 
@@ -34,7 +34,7 @@ module Search
       end
 
       test '#delete_index! deletes an index' do
-        index = GradeIndex.new
+        index = IdentityIndex.new
         index.create_index!
         assert index.index_exists?
 
@@ -43,7 +43,7 @@ module Search
       end
 
       test "defines mappings and settings" do
-        index = GradeIndex.new
+        index = IdentityIndex.new
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_mappings").get.body)
         assert_kind_of Hash, resp[index.index_name]['mappings'][index.type_name]
 
@@ -52,16 +52,15 @@ module Search
       end
 
       test "#serializer points to the corresponding model ActiveModel::Serializer" do
-        index = GradeIndex.new
-        assert_equal GradeSerializer, index.serializer
+        index = IdentityIndex.new
+        assert_equal IdentitySerializer, index.serializer
       end
 
       test '#save' do
-        index = GradeIndex.new
+        index = IdentityIndex.new
         name = SecureRandom.hex(8)
-        grade = Grade.create name: name, review_status: ReviewStatus.reviewed
-
-        index.save(grade)
+        identity = Identity.create name: name
+        index.save(identity)
         refresh_indices
 
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_search?q=name:#{name}").get.body)
@@ -70,28 +69,28 @@ module Search
       end
 
       test '#after_save' do
-        index = GradeIndex.new
-        grade = grades(:grade_1)
+        index = IdentityIndex.new
+        identity = identities(:khan)
 
         def index.after_save(obj, res)
           @after_save_called = true
         end
 
-        index.save(grade)
+        index.save(identity)
         assert index.instance_variable_get(:@after_save_called)
       end
 
       test '#delete' do
-        index = GradeIndex.new
+        index = IdentityIndex.new
         name = SecureRandom.hex(8)
-        grade = Grade.create name: name
-        index.save(grade)
+        identity = Identity.create name: name
+        index.save(identity)
         refresh_indices
 
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_search?q=name:#{name}").get.body)
         assert_equal 1, resp['hits']['total']
 
-        assert index.delete(grade)
+        assert index.delete(identity)
         refresh_indices
 
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_search?q=name:#{name}").get.body)
@@ -99,10 +98,10 @@ module Search
       end
 
       test '#after_delete' do
-        index = GradeIndex.new
+        index = IdentityIndex.new
         name = SecureRandom.hex(8)
-        grade = Grade.create name: name
-        index.save(grade)
+        identity = Identity.create name: name
+        index.save(identity)
         refresh_indices
 
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_search?q=name:#{name}").get.body)
@@ -112,7 +111,7 @@ module Search
           @after_delete_called = true
         end
 
-        assert index.delete(grade)
+        assert index.delete(identity)
         assert index.instance_variable_get(:@after_delete_called)
       end
     end
