+++ test/models/search/indices/resource_type_index_test.rb
@@ -2,12 +2,12 @@ require 'securerandom'
 require 'test_helper'
 
 module Search
-  module Indexes
-    class StandardIndexTest < ActiveSupport::TestCase
+  module Indices
+    class ResourceTypeIndexTest < ActiveSupport::TestCase
       @@initial_setup = true
 
       setup do
-        index = StandardIndex.new
+        index = ResourceTypeIndex.new
         if @@initial_setup || !index.index_exists?
           index.reset_index!
           @@initial_setup = false
@@ -15,17 +15,17 @@ module Search
       end
 
       test "#type_name" do
-        index = StandardIndex.new
-        assert_equal 'standard', index.type_name
+        index = ResourceTypeIndex.new
+        assert_equal 'resource_type', index.type_name
       end
 
       test "#index_name is composed with type and environment" do
-        index = StandardIndex.new
-        assert_equal 'standards__test', index.index_name
+        index = ResourceTypeIndex.new
+        assert_equal 'resource_types__test', index.index_name
       end
 
       test '#create_index! creates a new index' do
-        index = StandardIndex.new
+        index = ResourceTypeIndex.new
         index.delete_index!
         refute index.index_exists?
 
@@ -34,7 +34,7 @@ module Search
       end
 
       test '#delete_index! deletes an index' do
-        index = StandardIndex.new
+        index = ResourceTypeIndex.new
         index.create_index!
         assert index.index_exists?
 
@@ -43,7 +43,7 @@ module Search
       end
 
       test "defines mappings and settings" do
-        index = StandardIndex.new
+        index = ResourceTypeIndex.new
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_mappings").get.body)
         assert_kind_of Hash, resp[index.index_name]['mappings'][index.type_name]
 
@@ -52,15 +52,15 @@ module Search
       end
 
       test "#serializer points to the corresponding model ActiveModel::Serializer" do
-        index = StandardIndex.new
-        assert_equal StandardSerializer, index.serializer
+        index = ResourceTypeIndex.new
+        assert_equal ResourceTypeSerializer, index.serializer
       end
 
       test '#save' do
-        index = StandardIndex.new
+        index = ResourceTypeIndex.new
         name = SecureRandom.hex(8)
-        standard = Standard.create name: name
-        index.save(standard)
+        resource_type = ResourceType.create name: name
+        index.save(resource_type)
         refresh_indices
 
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_search?q=name:#{name}").get.body)
@@ -69,28 +69,28 @@ module Search
       end
 
       test '#after_save' do
-        index = StandardIndex.new
-        standard = standards(:ccls_1_2)
+        index = ResourceTypeIndex.new
+        resource_type = resource_types(:lesson)
 
         def index.after_save(obj, res)
           @after_save_called = true
         end
 
-        index.save(standard)
+        index.save(resource_type)
         assert index.instance_variable_get(:@after_save_called)
       end
 
       test '#delete' do
-        index = StandardIndex.new
+        index = ResourceTypeIndex.new
         name = SecureRandom.hex(8)
-        standard = Standard.create name: name
-        index.save(standard)
+        resource_type = ResourceType.create name: name
+        index.save(resource_type)
         refresh_indices
 
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_search?q=name:#{name}").get.body)
         assert_equal 1, resp['hits']['total']
 
-        assert index.delete(standard)
+        assert index.delete(resource_type)
         refresh_indices
 
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_search?q=name:#{name}").get.body)
@@ -98,10 +98,10 @@ module Search
       end
 
       test '#after_delete' do
-        index = StandardIndex.new
+        index = ResourceTypeIndex.new
         name = SecureRandom.hex(8)
-        standard = Standard.create name: name
-        index.save(standard)
+        resource_type = ResourceType.create name: name
+        index.save(resource_type)
         refresh_indices
 
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_search?q=name:#{name}").get.body)
@@ -111,7 +111,7 @@ module Search
           @after_delete_called = true
         end
 
-        assert index.delete(standard)
+        assert index.delete(resource_type)
         assert index.instance_variable_get(:@after_delete_called)
       end
     end
