+++ test/models/search/indices/standard_index_test.rb
@@ -2,12 +2,12 @@ require 'securerandom'
 require 'test_helper'
 
 module Search
-  module Indexes
-    class LanguageIndexTest < ActiveSupport::TestCase
+  module Indices
+    class StandardIndexTest < ActiveSupport::TestCase
       @@initial_setup = true
 
       setup do
-        index = LanguageIndex.new
+        index = StandardIndex.new
         if @@initial_setup || !index.index_exists?
           index.reset_index!
           @@initial_setup = false
@@ -15,17 +15,17 @@ module Search
       end
 
       test "#type_name" do
-        index = LanguageIndex.new
-        assert_equal 'language', index.type_name
+        index = StandardIndex.new
+        assert_equal 'standard', index.type_name
       end
 
       test "#index_name is composed with type and environment" do
-        index = LanguageIndex.new
-        assert_equal 'languages__test', index.index_name
+        index = StandardIndex.new
+        assert_equal 'standards__test', index.index_name
       end
 
       test '#create_index! creates a new index' do
-        index = LanguageIndex.new
+        index = StandardIndex.new
         index.delete_index!
         refute index.index_exists?
 
@@ -34,7 +34,7 @@ module Search
       end
 
       test '#delete_index! deletes an index' do
-        index = LanguageIndex.new
+        index = StandardIndex.new
         index.create_index!
         assert index.index_exists?
 
@@ -43,7 +43,7 @@ module Search
       end
 
       test "defines mappings and settings" do
-        index = LanguageIndex.new
+        index = StandardIndex.new
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_mappings").get.body)
         assert_kind_of Hash, resp[index.index_name]['mappings'][index.type_name]
 
@@ -52,15 +52,15 @@ module Search
       end
 
       test "#serializer points to the corresponding model ActiveModel::Serializer" do
-        index = LanguageIndex.new
-        assert_equal LanguageSerializer, index.serializer
+        index = StandardIndex.new
+        assert_equal StandardSerializer, index.serializer
       end
 
       test '#save' do
-        index = LanguageIndex.new
+        index = StandardIndex.new
         name = SecureRandom.hex(8)
-        language = Language.create name: name
-        index.save(language)
+        standard = Standard.create name: name
+        index.save(standard)
         refresh_indices
 
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_search?q=name:#{name}").get.body)
@@ -69,28 +69,28 @@ module Search
       end
 
       test '#after_save' do
-        index = LanguageIndex.new
-        language = languages(:en)
+        index = StandardIndex.new
+        standard = standards(:ccls_1_2)
 
         def index.after_save(obj, res)
           @after_save_called = true
         end
 
-        index.save(language)
+        index.save(standard)
         assert index.instance_variable_get(:@after_save_called)
       end
 
       test '#delete' do
-        index = LanguageIndex.new
+        index = StandardIndex.new
         name = SecureRandom.hex(8)
-        language = Language.create name: name
-        index.save(language)
+        standard = Standard.create name: name
+        index.save(standard)
         refresh_indices
 
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_search?q=name:#{name}").get.body)
         assert_equal 1, resp['hits']['total']
 
-        assert index.delete(language)
+        assert index.delete(standard)
         refresh_indices
 
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_search?q=name:#{name}").get.body)
@@ -98,10 +98,10 @@ module Search
       end
 
       test '#after_delete' do
-        index = LanguageIndex.new
+        index = StandardIndex.new
         name = SecureRandom.hex(8)
-        language = Language.create name: name
-        index.save(language)
+        standard = Standard.create name: name
+        index.save(standard)
         refresh_indices
 
         resp = JSON.parse(Faraday.new(:url => "#{es_url}/#{index.index_name}/_search?q=name:#{name}").get.body)
@@ -111,7 +111,7 @@ module Search
           @after_delete_called = true
         end
 
-        assert index.delete(language)
+        assert index.delete(standard)
         assert index.instance_variable_get(:@after_delete_called)
       end
     end
